[#1] Improving UI for Users Dashboard for v1.0 - Abstracting dashboard styling and structure to CsuDashboard (should be moved to CSU Template) - Adding GroupListItem to display Group description in EditUser
