+++ src/testData/users.json
@@ -1,0 +1,62 @@
+[{
+  "eId": 0,
+  "eName": "ename",
+  "first_name": "first_name",
+  "last_name": "last_name",
+  "email": "email",
+  "profile_image": "profile_image",
+  "last_active": "2017-07-13T15:49:29Z",
+  "userGroups": [{
+    "user_group_type_id": 0,
+    "name": "name",
+    "alias": "alias",
+    "description": "description00"
+  }, {
+    "user_group_type_id": 1,
+    "name": "name",
+    "alias": "alias",
+    "description": "description01"
+  }]
+}, {
+  "eId": 1,
+  "eName": "ename",
+  "first_name": "first_name",
+  "last_name": "last_name",
+  "email": "email",
+  "profile_image": "profile_image",
+  "last_active": "2017-07-13T15:49:29Z",
+  "userGroups": [{
+    "user_group_type_id": 0,
+    "name": "name",
+    "alias": "alias",
+    "description": "A very long description. A very long description. A very long description. A very long description. A very long description. "
+  }]
+}, {
+  "eId": 2,
+  "eName": "ename",
+  "first_name": "first_name",
+  "last_name": "last_name",
+  "email": "email",
+  "profile_image": "profile_image",
+  "last_active": "2017-07-13T15:49:29Z",
+  "userGroups": [{
+    "user_group_type_id": 0,
+    "name": "name",
+    "alias": "alias",
+    "description": "description20"
+  }]
+}, {
+  "eId": 3,
+  "eName": "ename",
+  "first_name": "first_name",
+  "last_name": "last_name",
+  "email": "email",
+  "profile_image": "profile_image",
+  "last_active": "2017-07-13T15:49:29Z",
+  "userGroups": [{
+    "user_group_type_id": 0,
+    "name": "name",
+    "alias": "alias",
+    "description": "description30"
+  }]
+}]
